+++ test/UserAgentBag.test.js
@@ -1,4 +1,3 @@
-// @flow
 const test = require('ava')
 const UserAgentBag = require('../UserAgentBag')