+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Removed
 
+- Dropped support for old Node versions. Node v14+ is required
 - Removed Flow support
 
 ## 0.2.0 - 2019-05-16